WebGPU: refactoring of upload and dynamic memory managers

+++ Graphics/GraphicsEngineWebGPU/include/BufferWebGPUImpl.hpp
@@ -99,7 +99,7 @@ public:
 
     void Unmap();
 
-    Uint64 GetAlignment() const;
+    Uint32 GetAlignment() const;
 
     const DynamicMemoryManagerWebGPU::Allocation& GetDynamicAllocation(DeviceContextIndex CtxId) const;
 
@@ -132,7 +132,7 @@ private:
 
     WebGPUBufferWrapper   m_wgpuBuffer;
     DynamicAllocationList m_DynamicAllocations;
-    const Uint64          m_Alignment;
+    const Uint32          m_Alignment;
 };
 
 } // namespace Diligent
